% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{duration_custom}
\alias{duration_custom}
\title{Specify a custom distribution for the duration of positivity}
\usage{
duration_custom(pmf)
}
\arguments{
\item{pmf}{Probability mass function (vector) for the distribution (automatically normalised), assumed to have support 1:length(pmf)}
}
\description{
Specify a custom distribution for the duration of positivity
}
