% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{infections_flat}
\alias{infections_flat}
\title{A flat epidemic}
\usage{
infections_flat(possible_times = 1:100)
}
\arguments{
\item{possible_times}{a vector of times at which infections can occur}
}
\description{
A flat epidemic
}
