% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sensitivity_fixed}
\alias{sensitivity_fixed}
\title{A fixed test sensitivity with no false positives}
\usage{
sensitivity_fixed(sensitivity = 1)
}
\arguments{
\item{sensitivity}{The test sensitivity to use, the default is no false negatives}
}
\description{
A fixed test sensitivity with no false positives
}
