% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_cis}
\alias{simulate_cis}
\title{Simulate the CIS study}
\usage{
simulate_cis(
  positive_duration,
  n_indivs = 1e+06,
  testing_times = testing_fixed(),
  infection_times = infections_flat(),
  testing = sensitivity_fixed()
)
}
\arguments{
\item{positive_duration}{A duration function specifying the duration distribution to use}

\item{n_indivs}{The number of individuals (episodes) to simulate}

\item{testing_times}{A testing function specifying the times at which individuals are tested}

\item{infection_times}{An infections function specifying the shape of the epidemic}

\item{testing}{A sensitivity or other function specifying how to simulate the testing results}
}
\description{
Simulate the CIS study
}
