% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{simulation_to_episodes}
\alias{simulation_to_episodes}
\title{Summarise a simulation as episodes}
\usage{
simulation_to_episodes(tbl_sims)
}
\arguments{
\item{tbl_sims}{A tibble of tests as returned by positive_duration}
}
\description{
Summarise a simulation as episodes
}
