% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{testing_custom}
\alias{testing_custom}
\title{Custom testing times}
\usage{
testing_custom(tbl_times)
}
\arguments{
\item{tbl_times}{Tibble with two columns: individual and test_time for each test}
}
\description{
Custom testing times
}
