% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{testing_fixed}
\alias{testing_fixed}
\title{Fixed interval testing times}
\usage{
testing_fixed(test_interval = 7, first_test = -90, last_test = 300)
}
\arguments{
\item{test_interval}{Time between tests.}

\item{first_test}{First possible testing time. Each individual will be tested first at a time between `first_test` and `first_test + test_interval - 1` chosen uniformly at random (this gives an even spread of testing times).}

\item{last_test}{Last possible testing time.}
}
\description{
Fixed interval testing times
}
